requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
docker
psycopg2-binary
asyncpg
docker
python-dotenv
pydantic_settings
psycopg2-binary